+++ NEWS.md
@@ -1,3 +1,8 @@
+## Version 0.99.1 (2026-01-19)
+
+* `local_offset_distance_with_background` can now handle non-Xenium transcript
+data with no QV column without data modification from user.
+
 ## Version 0.99.0 (2025-12-01)
 
 * Submitted to Bioconductor